Move the item mapper to its own class
